+++ src/features/popularMovies/moviesSaga.js
@@ -14,7 +14,7 @@ import {
   fetchSearchMovies,
 } from "./moviesSlice";
 import { getPopularMovies } from "./getPopularMovies";
-import { getGenresList } from "../getGenresList";
+import { getGenresList } from "../../getGenresList";
 import { getSearchResults } from "../../common/Navigation/Search/getSearchResults";
 
 function* fetchPopularMoviesHandler({ payload: page }) {
